[ion] Increase the stack size to 32 KB
Turns out 4K was too small for our usage pattern (we measured an
overflow when drawing function curve - probably because we store a
little bit of bitmap data on the stack). So the same area was used for
the heap *and* the stack, causing terrible memory corruption.

We could use a stronger pattern (stack on the SRAM boundary + MPU to
trigger an exception when overflowing the stack) to prevent memory
smashing.

Change-Id: I4d187e154322f18e1e7deddd75f7242af777ade3