[ion] Toggle between AlphaLock and ShiftAlphaLock
Change the behaviour of the 'shift' key in the AlphaLock and
ShiftAlphaLock states to toggle between the two.  This makes it easy to
switch between upper case and lower case letters in ALPHA-lock mode,
although the common case of entering words with isolated captital
letters in alpha-lock mode now involves additional presses of the
'shift' key.

With this change, it is impossible to enter the AlphaLockShift state.