[build] Run python2 explicitly in apps/Makefile
The "apps/i18n.py" script is not compatible with Python 3.
"apps/Makefile" uses the 'python' executable to run the script.  On some
systems (such as Arch Linux), the 'python' executable is a version of
Python 3, which breaks the build.  Follow the recommendation of PEP 394
and run the 'python2' command instead.  Also change the "shebang" line
of the script to refer to python2 although we do not rely on the shebang
line to run python.

There might be some systems without 'python2' where 'python' is Python
2.  This fix will break the build on those systems, but there should be
an easy workaround.