adjust German wording in Settings app
I propose changing EditionMode from "Eintrag" to "Eingabe", as the latter is more commonly used in German to describe the entry of text in a device. Also, "Linear" is a more apt description of "In Zeilen" (the latter suggests there are always multiple lines involved, which isn't strictly the case). Finally, "Natürlich" got rid of the trailing "e", which isn't necessary here.