[liba] Update the bool type to use _Bool
_Bool is defined in C99 which we use already. In practice it's 1 byte
just like char, so this shouldn't have any impact on the built object.

But this patch removes LTO errors when linking C code with C++ code.
Indeed, the linker would raise type mismatch because in C++ bool is
defined to _Bool whereas in C it would be defined to char.