[poincare/layout_cursor] Fix cursor position after adding square root
Scenario:
Go to the 'Calculation' app
Type 'sqrt(sqrt(i))', by pressing the square button twice and then the i button
Press OK
Scroll up, and press OK
Press left until the cursor is right before the i.
Press "sqrt" : the root is added, but the cursor jumped to the right of the i.
The cursor should have stayed on the left of the i.

Note: The behavior is correct when you type the entire expression straight away by doing the following:
Press "sqrt", "sqrt", "i"
Press "left", so the cursor is right before the i
Press "sqrt" again. At this point, the cursor remains on the left of i.