[poincare] Prevent unnecessary errors of mathematical functions
This removes the two functions Trigonometry::RoundToMeaningfulDigits and
ApproximationHelper::TruncateRealOrImaginaryPartAccordingToArgument (and
their uses) which are used for rounding the results of almost all
mathematical functions exposed to the user on the calculator. Also
removed is a similar procedure used inline after the computation of
derivatives.

The intent of the removed code was to round the results of mathematical
functions with the supposed accuracy of the underlying liba (OpenBSD
libm) and libaxx (libcxx) functions in mind, so no significant error
would be introduced.

The functions have their implementation and interface issues: they can
create a much larger error of the result of the mathematical function
than intended and fixing them to work as intended would necessitate a
lot of work and increasing code size, as each mathematical function
would need its own "rounding function" to prevent errors greater than
intended.

However, even if the removed functions were made to work as intended at
whatever cost, their use in Poincare in the current manner could still
produce horrible results. This is because even insignificant errors can,
with propagating the rounded outputs of functions to other functions
which again round their outputs, ruin the total result at the end of the
calculation (because of a greatly magnified error).

Also note that the problem the "rounding functions" were intended to
solve can be solved in a simpler manner by giving the user the (perhaps
default) option of a fixed point fractional number display (like the %f
conversion specifier in C printf), coupled with more usage of tables of
notable values of mathematical functions.

Fixes #1373

Updates #1350 (The left part of the plot with the imaginary part is
gone, but the right part's plot still has the steps/staircase. The
values shown for each x are accurate and do not correspond to the
staircase, though.)

Updates #1222 (The plot is a lot nicer, but still has some gaps.)