[poincare/derivative] Avoid creating Derivative nodes in method derivate
When a derived expression contains other derivatives, the derivate
method will attempt to call derivate on it instead of building a
Derivative node and relying on its reduction. If the expression cannot
be derived formally, an irreducible Derivative node is created anyway.
This prevents Dependency nodes from being added at each step of the
derivation process.