[shared/sequence_store] Method recordAtNameIndex
When handling sequences, the method `recordAtIndex` was used with the
name index as argument (0 for u, 1 for v, 2 for w).
However, if the sequences are defined in another order, recordAtIndex(0)
will return the first defined sequence, which may not be u.
The method recordAtNameIndex covers the use cases in which recordAtIndex
was mistakenly used before.