[solver] Reset memoization on record change
Unlike Graph and Sequence, the Solver did not notify the EquationStore
to reset its memoization when the storage was tampered with. As such,
the memoized equations were allowed to contain data about deleted records.

However, this was not an issue, as creating a new equation always
started with injecting new text into the record, which deleted the old
content. Now, with the possibility to create an equation right from the
list without using the type selector, the user was able to resurrect
equations that had just been deleted.

e.g.:
  - In Solver, with no equations, define "x=0"
  - Select the equation and press Backspace to delete it
  - Type "1" on the "Add equation" button to start a new equation
--> The edition field displays "1", but the equation field at the top of
the screen displays "x=0" which was previously deleted.