[function_graph_controller] Fix range ratio
To avoid hiding curves like x^2 behing the graph banner, we optimize the
display for a shorter range. To do that, we used to consider a range of
ratio N / (1 + m), where N is the orthonormal ratio and m the combined
vertical margins.
However, based on how the margins are added later on, the correct ratio
was N * (1 - m).