[graph, regression, sequence] Pop until graph controller
Returning to the graph from a submenu in the OK menu used to be done by
chaining several stack->pop().
Now that the OK menu can be reached either by pressing OK on a curve, a
via the curve selector, the depth of the submenu is not constant.
To come back to the graph, we use popUntilDepth(1), as the graph
controller is always at depth 1 no matter the path.