[calculation/calculation] Allow output expression to be uninitialized
Output expressions are built by parsing the output text. If the pool is
full, it can result in an uninitialized expression. As such, we remove
the assertion that those expressions must be initialized.
It is up to the caller of input(), exactOutput() and approximateOutput()
to determine the right course of action if the expression is
uninitialized.