[statistics/calculation_controller] Implement MarginDelegate
The cells used in CalculationController have an inner separator, so that
there seems to be a margin between contiguous cells. When the first
series is selected, this separator would combine with the preface
margin, and would look like a double margin. As such, we deactivate the
margin in this case.