[poincare/code_point_layout] Change handling of thousand separation
Instead of declaring a block of code points as Integers and finding the
position of the separators when displaying the layout, the code points
now have their types set when building the layout, to speed up the
display. This relies on the fact that display types are not used in
edition.