[ion/kernel/keyboard] Queue state as a callback from scan
To properly scan the keyboard, we need to shut down the EXTI lines.
However, this means a python script with a 'keydown' loop will spend
most of its time with the keyboard interruptions turned off, and as such
will be hard to stop.
To prevent that, instead of queueing the event returned by 'scan' in the
'poll' method, we queue all state read by 'scan' directly in the 'scan'
method. This way, interrupting keyboard states read by python will be
present in the queue when python checks whether it should stop the
execution.