[poincare/power] Special case for reduction of constant i
With reduction targets SystemForApproximation and SystemForAnalysis, the
constant i is not transformed into a ComplexCartesian ; as such, most
reduction rules do not apply.
Add a special rule to reduce the power of type i^(p/q), since those are
easy to compute and can greatly reduce the complexity of the expression
if p is even.