[poincare/power] Perform upper bound computation in double
We do not reduce a power of rationals if the numerator or denominator
would exceed FLT_MAX, otherwise it would cause some graphing glitches.
This check must be performed whit double precision computation. If the
computations uses float, then the computation of pow(10^40, 0.5) would
evaluate to +Inf (since 10^40 is +Inf in single precision), although its
result is representable in single precision.