[apps/graph] Zoom out to ensure an integral/sum bound is visible
Make sure the (x,y) and (x,0) pair is always visible by zooming out.
Note that the zoom ratio is purposedly not kept. Indeed, if we wanted to
keep the zoom ratio, then scrolling right on an integral bound could
result in the cursor actually moving left if the function was growing
faster on the y-axis than on the x-axis (e.g. exp(x)).