[escher/src] Simplify returnToRootMenu
It was calling returnToPreviousMenu which would, for toolboxes,
use the states to go back menu by menu. If returnToRootMenu
had been called because of an invalid checksum, this operation
could crash the calculator.
returnToRootMenu migth need to be overriden in
AlternateEmptyNestedMenuController if we want to use it in the
future as it might not handle empty views properly.