[apps][escher] App::willBecomeInactive doesn't need to dismiss modal views since the appp is about to be destructed anyway
This fixes (temporarily?) the following bug: we ran into a memory
exceptions. Some tree handles are "dangling" but their identifiers are
labelled "free". In the willBecomeInactive, we allocated a new handle
with a identifier colluding with the "dangling" one. When destructing
the application, we destruct the node pointed twice: once with the
rightfull handle and once by the dangling one (which can't detect its
dangling state because its identifier has been populated).: