[apps/probability] Add freedom degree selector
For goodness test, there is now an editable cell for the degree of freedom.
We have to un-factorize the InputCategoricalView class so that
they can have a different cells.

We also have to give the GoodnessTabelViewController a pointer to the
InputGoodnessView so it can update the degreeOfFreedom cell
when the table size change.