[simulator] Refresh screen on web python
The refresh was done in the scan, but commit 9bfa204 "[ion] Simulator and
device keyboard works similarly by queuing keyboard" changed the times
where we could call scan, so the refresh was no longer done at the right
moments. As a result, Python drawings on a web emulator would not be
displayed in real time.