[poincare/floor,ceil] Add tolerance around integers
The floor and ceiling functions can easily magnify a small approximation
error around an integer. If their argument is very close the an integer
(relative error of DBL_EPSILON), we assume the difference is more likely
to come from an approximation error than a user input, and treat the
value as the nearest integer.