[code] Fix position python heap
The python heap was extended to 64KiB by moving the pool right before
the Apps buffer. However, this caused the vtable for Code::App to be
located inside the heap. The pool is now after the Apps buffer, and the
python heap extension is the last member of Code::App.