[poincare/tests] Remove useless polynomial test
This use to test if the result was only approximated because it could
not be reduced.
Then it was decided that even if an expression could not be reduced, the
reduction would return the un-reduced expression.
From there, the test had no real purpose.

Now that cos^2+sin^2 is reduced, the result can be reduced but its
result is very complicated. Since the test has no purpose anymore and
is difficult to maintain, we decided to remove it.