[apps/code] displayEmptyControllerIfNeeded also when varbox becomes first responder
Linked to commit 863974b27b39e0341ebeaa46c67939dacc327534 : when root menu of VariableBox is empty, responder should not be SelectableTableView, so we switch to VariableBoxEmptyController to avoid it. The latter commit only did it when VariableBoxController did enter responder chain, but not when it did become first responder.

Indeed when we open varbox in python, ModalViewController::displayModalViewController calls setFirstResponder several times :
  1) in m_currentModalViewController->viewWillAppear() : which will later call setFirstResponder(&m_listController), which calls VariableBoxController->didEnterResponderChain() which calls displayEmptyControllerIfNeeded() --> perfect, that's what we want
  2) in Container::activeApp()->setFirstResponder(vc) : will only call VariableBoxController->didBecomeFirstResponder() but not VariableBoxController->didEnterResponderChain() since varbox is already in responder chain --> sad ... we still wanted to call displayEmptyControllerIfNeeded()