[shared/editable_cell_table_view_controller] Remove invalid assert
Display asserted that cells were not being edited (to avoid blinking on
the screen). This assertion relied on the fact that the table was never
fully redrawn while a cell was being edited. This assumption is now
invalid, as the store menu can be summoned from an edited cell, and can
trigger a full reload.