[escher/text_field] Set cursor location when setting text only if the text buffer is the edited text buffer
Previously to this change, the cursor could change position in
editedText() while the editedText buffer was not the one filled by
setText, which could result in the cursor being out of bounds or between
two chars of the same code point
This would happend when an invalid interval bound of a function was inputted while xAuto was True.
When textFieldDidFinishEditing was called, the xAuto is set to false,
which triggers a reload of the table, and set the text to its value
before editing.