[apps/shared] Separate intersection of row and column prefaces
In PrefacedTwiceTableView, the column preface used the full table view height with a hack when a preface row is displayed, namely artificially changes the first row displayed. This hack created a problem with the column preface offset which was no more consistent. To avoid this problem, the intersection of preface row and preface column is now separated from the preface colum).