[apps/graph] Merge classes FunctionParameterController and FunctionColumnParameterController
The class FunctionParameterController is only used throught an object of
type FunctionColumnParameterController, indeed only using the methods
from the latter class. The controllers are exactly the same.

Moreover we don't need the "column" indicator in the name of the class,
it is never done for classes inheriting from ColumnParameterController (cf
DerivativeParameterController, ValuesParameterController ...)