[apps/solver] Fix exact equal sign in solution
Fixes a bug that appeared with commit 2f9a141e66f904caedcb8f21f15c970b93676c6f.
After this commit, in solutions of app solver, there was an approximate equal sign when we needed an exact equal sign. For example: for equation x-1/2=0, solution was 1/2 approx= 0.5.

The reason was that before, we parsed the string of approximate solution to compare it with the exact solution. When parsing, the approximation solution expression was of type Decimal, which is equivalent to a Rational. Whereas exactSolutionsApproximations is an expression of type Float, which is not equivalent to a Rational.