[poincare] Fix fuzzer crash about matrix that failed to be reduced
Scenario:
1. go in app grapher
2. choose template f(t)=[cos(t),sin(t)]
3. go at the end of sin(t) and add
   [tan(tan(tan(tan(tan(tan(tan(i*sin(5))))))))]
4. EXE

The problem was that when we tried to reduce this expression, the pool
got full and we failed to reduce. Later, when we tried to add a caption
for this function, we assumed that this matrix was valid because it was
reduced. The result of the reduction of a matrix containing
another matrix is an undef expression, so when reduction doesn't fail,
we don't even look for parametric captions. See for example with only 6
tan and not 7: pool doesn't get full, reduction succeed and the caption
is "Undefined".