[apps/shared] Create a method floatToPixelIndex that returns a KDCoordinate
Fixes issues #2664 and #3672
The problem was that we computed a pixel in float, then did some
computation between pixels and then rounded the result to have a KDCoordinate.
Depending on the order of the computations, results could be different. Now,
we compute every pixel in KDCoordinate directly and then do the calculus bewteen
pixels after.