[apps/shared] Replace m_borderWidth by a bool m_displayBorder for histograms
Indeed, currently in the code there is only 2 values possible for
borderWidth : 0 and 1. If borderWidth were to be > 1, the code might
not be consistent because we wouldn't know at which value to place it on
the axis. The code works well for border width 0 and 1, and I don't
think we will ever want to have a border > 1, so for now this is clearer
to add a bool m_displayBorder and a constexpr k_borderWidth.