[calculation/additional_outputs] Fix TrigonometryListController computing wrong angle because it was changing the angle unit
We used to convert e to radians to make calculations easier, but we did
it by multiplying it by pi/180 (in case of angleUnit == degrees) and
then setting angleUnit to radians. This method was producing wrong
results because if e = cos(2pi) for example (meaning that the user input
was cos(cos(2pi)), cos(2pi) in degrees is different from (pi/180) *
cos(2pi) in radians.
To ensure that all computations are properly done, we now do them in the
original angle unit, and convert to radians only at the end of the
method.