[apps/shared] Remove while loop in PrefacedTwiceTableView
While loops are dangerous, so I found a way around :
an abscissa column cannot be frozen if we can see both the abscissa
column and its last values column in the same screen frame.
That way, we avoid to try to fix this sort of abscissa column and then
realise we just cropped its last values column, which means we have to
change the column to freeze (reason why I added the while loop).