[apps/shared] Fix visual artifact
Scenario grapher_values_preface.nws showed a visual artifact: when
changing the size of a cell, the prefaced table changes its offset.

Thus we now avoid changing the value of the offset when deselecting
temporarily the table. We already did it before but removed it here: https://github.com/numworks/epsilon-internal/pull/3616
because at that time, it created another visual artifact. I thus added
a scenario for this other artifact, that does not happen now.