[apps/solver] Fix fuzzer crash
Scenario: in app solver, enter "0=0 or x=0" and then solve.
We should have prevented before this type of equation, returning a
syntax error. The problem is that this expression delegate is shared
with interval expression delegate, and we can't accept only expressions
of type BinaryEquality. Currently, only apps can override isAcceptableExpression,
this should be refactored.