[poincare] Do not go up to derivand when at abscissa layout because
Indeed, when at abscissa layout, if we go up it is very likely that we
want to select previous computation instead of going to derivand. It can
be very long otherwise : if derivand is a multiple fraction, we need to
go multiple times Up to select previous computation in history.