[poincare] Fix replace symbol
When replacing u(n) for n for cobweb, we also replace u(0) for 0, which
we don't want. Currently, we never ask to replace f(3) or
u(4), but we always ask to replace f(x) or u(n) with x and n symbols.

If one day we need to ask to replace f(3) by 2x, we will need to change
this code and make the case disjunction before: we will need to ask to
replace u(n) by n only in the case n is a symbol.