[apps/inference/statistic/chi_square_and_slope] Don't save selected cell in results homogeneity controller
Fix #4295

If we wanted to fix this issue but save selected cell, we could have
added
m_resultHomogeneityTable.selectableTableView()->unhighlightSelectedCell()
and m_selectableListView.unhighlightSelectedCell() right after reload in
ResultsHomogeneityController::ResultsTableController::viewWillAppear.
Indeed, reloadData highlights the selected cell (both the selected cell
in list table view and the selected cell in categorical table cell). Sometimes we
don't want anything highlighted, for example when we change tab but do not
enter tab. In case we want to highlight something, it will be done when
setting responders.