[apps/regression & statistics] Useless override of tableViewDidChangeSelection
Already handled in indexOfNextSelectableColumn/Row: indeed, the 2 top
left cells are not selectable since m_state=Hidden

We still need to override cellAtLocationIsSelectable because with
prefaced tables, there is sometimes no margins when we ask for a cell
out of "view", it returns nulltpr, which is considered as selectable.

This does not happen without preface because with margins, we see a bit
of the cell we want to select so it "exists" already in the view.