[graph] Fix use of uninitialized variable
m_useColumnTitle, that would then force the use of
another uninitialized variable : m_titleBuffer in title().

Now it has a default value of false to ensure m_titleBuffer is not
used when uninitialized.
It is also properly set up when controller is opened.

The crash: Go to finance, then home, then create a function, plot it
and check it option via the graph menu.