[apps/calculation] Check first exact output (and not input) for trigo additional results
Fix fuzzer crash:
go to additional results of cos(binomial(0, -1%%%%%%%+200))

The problem was that we said "ok" for trigo additional results with this
input, but we took the exact output as expression for the additional results.

- With the input, we only know that -1%%%%%%%+200 is approximatively equal
to 1, so the binomial is 0 --> angle for trigo additional results is defined.

- With the exact output, we know that -1%%%%%%%+200 is -10^(-14)+200 (thanks
to percent beautification) which is not and int so the binomial is undef -->
angle for trigo additional results is not defined.