[apps] Set a flag in SequenceStore and ContinuousFunctionStore when the storage is changed
When an app is exited because of an Exception, the records of the
functions or sequences might need to be deleted. If not, the
Graph/Sequence app might be impossible to re-open. But since we do not
want to delete all functions/sequences each time an exception is raised,
they are deleted only if at least 1 function/sequence was edited since
the opening of the app.