[escher] Delete useless and confonding method SelectableListViewDataSource::selectCell
Indeed, it does exactly what does selectRow, and this latter name is way
clearer.

SelectableListView have a method selectCell and it does all the
intelligence of a selection: handle highlight and scroll, while SelectableListViewDataSource::selectCell
only change the member variable m_selectedRow. When we call selectCell,
we would expect some king of intelligence with cells, that's why it's
better to remove SelectableListViewDataSource::selectCell.