[poincare] Fix reduction when combining complex
An easiest way would have been to deepReduce the combination at the end
of combineComplexCartesians, but it is better to avoid deepReduce in
shallowReduce.

This fixes 2×cos(π/12)×e^(5πi/12) being reduced to 1/2+i(4+2√(3))/4
instead of 1/2+i(1+√(3))/2.

The reason was that during reduction, we had to add complex
c1 = (-1+√(3))/4 + i(1+√(3))/4
and complex
c2 = ( 3-√(3))/4 + i(3+√(3))/4
Both complex were reduced with reduction target User, so every real or
imaginary part was factorized under common denominator.
Then when adding the to complexes, we factorized everything under common
denominator like this:
c1 + c2 = (-1+√(3)+3-√(3))/4 + i(1+√(3)+3+√(3))/4 = 1/2 + i(4+2√(3))/4
The problem was that to simplify the imaginary part, we have to
distribute the addition, which is prohibited in after a factorisation on
common denominator to avoid infinite loops. Thus, the fraction was not
simplified, the shallow reduce did nothing.

However, if complexes were reduced with target analysis, ie
c1 = -1/4+√(3)\4 + i(1/4+√(3)/4)
and complex
c2 =  3/4-√(3)\4 + i(3/4+√(3)/4)
Then the addition of both would give directly
c1 + c2 = 1/2 + i(1+√(3)/2)
which will be shallow reduced in
c1 + c2 = 1/2 + i(2+√(3))/2