[build/screenshots] Delete useless script generate.py
This script was maybe used before to generate a folder of images from a
folder of state files. However now, we have another structure of folders
for the screenshot dataset: one folder for each scenario, containing the
state file and the image. Therefore the script generate.py is not very
helpful. Moreover, we now have the script add_scenario.py and
update_images.py.