[apps/calculation] No need to setDisplayCenter twice
We need to make updateExpanded void because by default m_calculationExpanded = false,
so when we set expanded to false, it won't recompute m_calculationExpanded and
setDisplayCenter in updateExpanded.

It is not a problem to reload scroll if expanded didn't change, anyway
we are setting a computation in a cell, so we might want to reload
scroll, even if the computation is the same.