[poinare/expression] Fix Expression::isReal for RandInt
Ensure Expression::isReal handles RandInt correctly. RandInt(a,b) is
considered real if both 'a' and 'b' are real numbers. Additionally,
it can be a list if either 'a' or 'b' is a list.
It fixes a crash when typing the following sequence in the calculation
app: randint(randint(randintnorep(0,0,0)×i,0)×i,0)×_C