[poincare/range] Enhance the algorithm of setMin/setMax
Calling setMin and then setMax could have unexpected results:

Lets say we want to create the range [9999,10000], with a limit of 10000
- We first create a NAN range [NAN,NAN]
- Then we set the min to 9999.
  > While setting the min, since the max is NAN, it is also set to 9999,
    but the distance between min and max in below k_minLength. So we
    shift the max away from by a value DefaultLength, which here is 100.
    But since the limit is 9999, the max couldn't be shifted enough and we had to
    shift the min instead, which gives [9899,9999] as a result.
- We finally set the max to 10000
- The result given is [9899, 10000] which is not what we wanted, while
  the range we wanted was perfectly valid.

With the new method, this shouldn't happen again, because when the max
is shifted to 10000, we recheck if the length is big enough. If it is,
we don't shift the min.