[poincare/power] Enhance power reduction
When reducing (a*b*c*d*)^e, we use to transform it into:
a^e*(b^e*(c^e*d^e)) which was calling shallowReduce an unecessary amount
of time. Now it's directly transformed into a^e*b^e*c^e*d^e

This solves a crash where a bad access was engendrated by the explosion
of the pool. I'm not sure exactly where the segfault was coming from but
this fixes it.