[apps/shared/continuous_function] Add condition in ContinuousFunction::Model::expressionEquation
The only "isValidNamedLeftExpression" with an equal sign not Equal is
the form f(x)<...

Thanks to previous commit, the form f(x)<... is already changed to
f(" ")<... if f already exists, or f*(" ")<... if f doesn't exist.
In both cases, f(x)<... won't be a "isValidNamedLeftExpression". So
adding the condition equationType == ComparisonNode::OperatorType::Equal
doesn't change anything (it just maked the code clearer, and will be
needed for factorisation in next commits).